Finish the index overflow/width sweep: tag CSR sum + doc-order ordinal
Two remaining spots brought in line with the text-index policy:

- attr_owner's tag CSR prefix sum (tag_off[t+1] = tag_off[t] + counts[t]) was
  unchecked. The total is n_indexed (bounded), but route it through mkr_size_add
  for uniformity, failing closed (LXB_STATUS_ERROR_MEMORY_ALLOCATION) on overflow.

- The XPath document-order index stored its ordinal as uint32_t, silently
  wrapping past 2^32 nodes. Widen ord to size_t (matching the text index's
  size_t ranges); the bucket struct stays 16 bytes (the uint32_t had 4 bytes of
  padding after it), so this is free, and the compare-don't-subtract logic is
  unchanged. The SplitMix bucket hash and the string-value cache's entry-index
  array keep their own (separate, bounded) uint32_t.

No behaviour change on reachable inputs. Verified: clean compile (no warnings),
C self-test true, security:clint 0 findings, spec 368/0/4 pending, ASan/UBSan
clean, fuzz:sanitize 0 findings, Nokogiri XPath differential 0 divergence /
0 order-only diff (ordering correct with the widened ordinal).

+++ ext/makiri/xpath/mkr_xpath_internal.h
@@ -363,7 +363,7 @@ void mkr_nodeset_unique_sorted(struct mkr_xpath_context_s *ctx, mkr_nodeset_t *n
 typedef struct {
   struct {
     const lxb_dom_node_t *node;  /* NULL = empty slot */
-    uint32_t              ord;
+    size_t                ord;   /* document-order ordinal (size_t: no 2^32 cap) */
   } *buckets;
   size_t cap;
   size_t count;
